requirements.txt
bitstring~=3.1.9
PyYAML~=6.0
bcoding==1.5
requests~=2.28.0
